simplify test codes
