+++ tests/exe_container_test.h
@@ -1,3 +1,4 @@
+#pragma once
 // Tests for json embedding
 // Todo: Write more tests
 
@@ -6,34 +7,7 @@
 #include "string_utils.h"
 #include "env_utils.h"
 
-const char* json_file;
-
-#ifdef _WIN32
-int wmain(int argc, wchar_t* argv[]) {
-#else
-int main(int argc, char* argv[]) {
-#endif
-    ::testing::InitGoogleTest(&argc, argv);
-    assert(argc == 2);
-
-#ifdef _WIN32
-    std::string argv1 = UTF16toUTF8(argv[1]);
-    json_file = &argv1[0];
-#else
-    json_file = argv[1];
-#endif
-
-    char *exe_dir = envuGetExecutableDir();
-    envuSetCwd(exe_dir);
-    envuFree(exe_dir);
-
-    return RUN_ALL_TESTS();
-}
-
-void GetTestJson(rapidjson::Document& json) {
-    json_utils::LoadJson(json_file, json);
-    EXPECT_FALSE(json.ObjectEmpty());
-}
+void GetTestJson(rapidjson::Document& json);
 
 TEST(JsonEmbeddingTest, Embed) {
     {
@@ -41,7 +15,7 @@ TEST(JsonEmbeddingTest, Embed) {
         test_json.SetObject();
         GetTestJson(test_json);
         ExeContainer exe;
-        json_utils::JsonResult result = exe.Read(json_file);
+        json_utils::JsonResult result = exe.Read(JSON_ALL_KEYS);
         EXPECT_TRUE(result.ok);
         exe.SetJson(test_json);
         result = exe.Write("embedded.json");
