+++ tests/json/config_ascii.json
